requirements.txt
# Основные зависимости для TripCraftBot
aiogram>=3.0.0,<4.0.0
redis>=5.0.0,<6.0.0
aiohttp>=3.8.0,<4.0.0
httpx>=0.25.0,<1.0.0
requests>=2.31.0,<3.0.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Зависимости для тестирования
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0

# Зависимости для разработки (опционально)
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.5.0,<2.0.0
ruff>=0.5.3,<1.0.0
types-redis>=4.6.0,<5.0.0
